Create Project3.java
Write Multicast client-server classes to do the following:
There are a number of clients. Each one send an integer to the server.
The server adds these integers.
When the server receives a zero from a client it displays the sum of the integers and terminates. 
For example if the first client sends 10, the second client sends 20, and the third client sends zero the server displays 30 and terminates.

To test your project make two folders one for the server and one for one client. Run the server first. Run the client several times and each time change the integer that the client is sending. In the last run of the client sends zero.

Sample Dialog

Clients:

Enter an integer: 10
Enter an integer: 20
Enter an integer: 0

Server:

The sum is: 30